requirements.txt
autopep8==1.4.4
beautifulsoup4==4.8.1
boto3==1.10.18
botocore==1.13.18
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==2.2.5
django-cors-headers==3.1.1
django-filter==2.2.0
django-rest-auth==0.9.5
django-storages==1.7.2
djangoql==0.13.1
djangorestframework==3.10.3
djangorestframework-datatables==0.5.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
entrypoints==0.3
flake8==3.7.8
future==0.18.1
ibm-cloud-sdk-core==1.0.0
ibm-watson==4.0.1
idna==2.8
install-scripts==0.1.6
jmespath==0.9.4
mccabe==0.6.1
moviepy==1.0.2
oauthlib==3.1.0
pep8==1.7.1
pexpect==4.7.0
ply==3.11
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-http-client==3.2.1
python-twitter==3.5
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
selenium==3.141.0
sendgrid==6.1.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
Strip==0.0.1.dev14
stripe==2.37.2
timeago==1.0.13
typing==3.7.4.1
tz==0.2.2
tzlocal==2.0.0
Unidecode==1.1.1
urllib3==1.25.6
uuid==1.30
watson-developer-cloud==2.10.1
websocket-client==0.48.0
xlrd==1.2.0
